visitEvent after close
